refactor(member): move security packages and update indentation to 2 spaces
