--- src/main/java/com/kindtalk/server/config/security/MyUserDetails.java
@@ -1,55 +1,0 @@
-package com.kindtalk.server.config.security;
-
-import com.kindtalk.server.member.domain.Member;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-
-@RequiredArgsConstructor
-public class MyUserDetails implements UserDetails {
-
-    private final Member member;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority("ROLE_" + member.getRole()));
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    @Override
-    public String getUsername() {
-        return member.getEmail();
-    }
-
-    @Override
-    public String getPassword() {
-        return member.getPassword();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
