--- src/main/java/com/kindtalk/server/config/security/dto/LoginResponse.java
@@ -1,8 +1,0 @@
-package com.kindtalk.server.config.security.dto;
-
-import com.kindtalk.server.member.role.Role;
-
-public record LoginResponse(
-    String email,
-    Role role) {
-}
