--- src/main/java/com/kindtalk/server/config/security/dto/LogoutResponse.java
@@ -1,5 +1,0 @@
-package com.kindtalk.server.config.security.dto;
-
-public record LogoutResponse(
-    String message) {
-}
