+++ src/main/java/com/kindtalk/server/security/dto/LoginResponse.java
@@ -1,0 +1,8 @@
+package com.kindtalk.server.security.dto;
+
+import com.kindtalk.server.member.role.Role;
+
+public record LoginResponse(
+  String email,
+  Role role) {
+}
