+++ src/main/java/com/kindtalk/server/security/dto/LogoutResponse.java
@@ -1,0 +1,5 @@
+package com.kindtalk.server.security.dto;
+
+public record LogoutResponse(
+  String message) {
+}
