+++ src/main/java/com/kindtalk/server/security/handler/CustomLogoutSuccessHandler.java
@@ -1,0 +1,31 @@
+package com.kindtalk.server.security.handler;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.kindtalk.server.security.dto.LogoutResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class CustomLogoutSuccessHandler implements LogoutSuccessHandler {
+
+  private final ObjectMapper objectMapper;
+
+  @Override
+  public void onLogoutSuccess(HttpServletRequest request,
+                              HttpServletResponse response,
+                              Authentication authentication) throws IOException {
+    response.setContentType("application/json;charset=UTF-8");
+    response.setStatus(HttpStatus.OK.value());
+
+    LogoutResponse res = new LogoutResponse("로그아웃되었습니다.");
+    response.getWriter().write(objectMapper.writeValueAsString(res));
+  }
+}
