+++ src/main/java/com/kindtalk/server/security/service/MyUserDetailsService.java
@@ -1,7 +1,8 @@
-package com.kindtalk.server.config.security;
+package com.kindtalk.server.security.service;
 
 import com.kindtalk.server.member.domain.Member;
 import com.kindtalk.server.member.repository.MemberRepository;
+import com.kindtalk.server.security.principal.MyUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -12,13 +13,13 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class MyUserDetailsService implements UserDetailsService {
 
-    private final MemberRepository memberRepository;
+  private final MemberRepository memberRepository;
 
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Member member = memberRepository.findByEmail(username)
-            .orElseThrow(() -> new UsernameNotFoundException("회원이 존재하지 않습니다."));
+  @Override
+  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+    Member member = memberRepository.findByEmail(username)
+      .orElseThrow(() -> new UsernameNotFoundException("회원이 존재하지 않습니다."));
 
-        return new MyUserDetails(member);
-    }
+    return new MyUserDetails(member);
+  }
 }
